refactor: move "email" package to node-app
