+++ imports/node-app/core-services/email/i18n/en.json
