+++ imports/node-app/core-services/email/i18n/fr.json
