+++ imports/node-app/core-services/email/i18n/sv.json
