+++ imports/node-app/core-services/email/index.js
@@ -1,6 +1,6 @@
-import i18n from "./i18n";
-import mutations from "./mutations";
-import startup from "./startup";
+import i18n from "./i18n/index.js";
+import mutations from "./mutations/index.js";
+import startup from "./startup.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
