+++ imports/node-app/core-services/email/mutations/index.js
@@ -1,0 +1,5 @@
+import sendEmail from "./sendEmail.js";
+
+export default {
+  sendEmail
+};
