--- imports/plugins/core/email/server/no-meteor/mutations/index.js
@@ -1,5 +1,0 @@
-import sendEmail from "./sendEmail";
-
-export default {
-  sendEmail
-};
